% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{age_prop}
\alias{age_prop}
\title{Age Proportions in the Population}
\format{
A numeric vector of length p, representing the age group proportions.
}
\usage{
age_prop
}
\description{
A vector representing the proportion of individuals in each age group for the population.
This is used for calculating immunity estimates for children and adults.
}
\keyword{datasets}
