% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{contact_matrix}
\alias{contact_matrix}
\title{Contact Matrix Between Age Groups}
\format{
A numeric matrix with dimensions [p x p] where rows and columns represent age groups.
}
\usage{
contact_matrix
}
\description{
A matrix representing the contact rates between different age groups in the population.
This matrix is used to estimate population immunity and disease transmission dynamics.
}
\keyword{datasets}
