% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_matrix.R
\name{fill_matrix}
\alias{fill_matrix}
\title{Fill a new matrix with values from an original matrix based on matching columns}
\usage{
fill_matrix(original_matrix, n_row, n_col)
}
\arguments{
\item{original_matrix}{A matrix whose values are transferred to the new matrix.}

\item{n_row}{An integer specifying the number of rows in the new matrix.}

\item{n_col}{An integer specifying the number of columns in the new matrix.}
}
\value{
A matrix with values from the original matrix placed in matching columns.
}
\description{
This function creates a new matrix of zeros with specified number of rows and columns,
then fills it with values from the original matrix based on matching column names.
}
