% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titer_related_funcs.R
\name{geo_mean}
\alias{geo_mean}
\title{Calculate the geometric mean of titer values}
\usage{
geo_mean(raw_titer)
}
\arguments{
\item{raw_titer}{A numeric vector of raw titer values}
}
\value{
The geometric mean of the titer values
}
\description{
This function calculates the geometric mean for raw titer values.
}
