% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuPop-package.R
\docType{package}
\name{ImmuPop-package}
\alias{ImmuPop}
\alias{ImmuPop-package}
\title{ImmuPop: Estimating Population Immunity Using Individual Antibody Titers}
\description{
Estimate population immunity metrics from individual serology data, including geometric mean titer (GMT), proportion of non-naive individuals, population immune proportion, and relative reduction in reproductive number (R0). Uses bootstrap sampling and MCMC-based Dirichlet estimation with age-structured contact matrices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/timktsang/ImmuPop}
  \item Report bugs at \url{https://github.com/timktsang/ImmuPop/issues}
}

}
\author{
\strong{Maintainer}: Tim Tsang \email{timkltsang@gmail.com}

Authors:
\itemize{
  \item Weijia Xiong \email{wjxiong@connect.hku.hk}
}

}
\keyword{internal}
