% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{ImmuPop_raw_data}
\alias{ImmuPop_raw_data}
\title{Example Data for Baseline Immunity}
\format{
A data frame with columns:
\describe{
  \item{uid}{Numeric, subject id.}
  \item{time}{Numeric, representing the time point.}
  \item{epi}{Character, epidemics}
  \item{bsl}{Character, indicating baseline status ("yes" or "no").}
  \item{raw_titer}{Numeric, raw titer values for the immunity estimation.}
  \item{age}{Numeric, age of subject.}
}
}
\usage{
ImmuPop_raw_data
}
\description{
A data frame that contains example data for baseline immunity estimation.
It includes information about time points, baseline status, age groups, and raw titers.
}
\keyword{datasets}
