% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply100}
\alias{multiply100}
\title{multiply 100 to the given number}
\usage{
multiply100(x)
}
\arguments{
\item{x}{a number, float or int}
}
\value{
a number
}
\description{
multiply 100 to the given number
}
\examples{
multiply100(5)
multiply100(-90)
}
