% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{protect_a}
\alias{protect_a}
\title{Protection Effect in the Population}
\format{
A numeric vector of length k, each represent protection effect from k-th antibody titer level,
for adult.
}
\usage{
protect_a
}
\description{
A vector representing the protection effect for children and adults.
This is used for immunity estimation.
}
\keyword{datasets}
