% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titer_related_funcs.R
\name{weighted_gmt}
\alias{weighted_gmt}
\title{Calculate the weighted geometric mean for the population}
\usage{
weighted_gmt(resampled_df, age_prop)
}
\arguments{
\item{resampled_df}{A dataframe containing raw titer values and age group information}

\item{age_prop}{A vector of age group proportions}
}
\value{
The weighted geometric mean for the population
}
\description{
This function calculates the weighted geometric mean across different age groups.
}
