% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuPop_allt_est.R
\name{ImmuPop_allt_est}
\alias{ImmuPop_allt_est}
\title{Estimate Immunity Across All Time Points}
\arguments{
\item{df_long}{A dataframe containing the data for all time points.}

\item{protect_c}{Numeric vector indicating the protection effect for children.}

\item{protect_a}{Numeric vector indicating the protection effect for adults.}

\item{agep_prop}{Numeric vector of age group proportions in the population.}

\item{contact_matrix}{Numeric matrix of contact rates between age groups.}

\item{sim_num}{The number of bootstrap simulations to run (default = 500).}
}
\value{
A dataframe with immunity estimates for each time point, including the median and 95% CI.
}
\description{
This function estimates immunity at each time point using `ImmuPop_timet_est` and returns the results.
}
