% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuPop_bsl_est.R
\name{ImmuPop_bsl_est}
\alias{ImmuPop_bsl_est}
\title{Estimate Immunity at the Baseline Period}
\arguments{
\item{df_long_bsl}{A dataframe containing baseline data (pre-epidemic), including epidemic information as `epi`.}

\item{protect_c}{Numeric vector indicating the protection effect for children.}

\item{protect_a}{Numeric vector indicating the protection effect for adults.}

\item{age_prop}{Numeric vector of age group proportions in the population.}

\item{contact_matrix}{Numeric matrix of contact rates between age groups.}

\item{sim_num}{The number of bootstrap simulations to run (default = 500).}
}
\value{
A dataframe with baseline immunity estimates, including the median and 95% CI.
}
\description{
This function estimates immunity at the baseline period for different epidemic (`epi`) and circulating virus groups.
}
