% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titer_related_funcs.R
\name{weighted_prop_HImorethan5}
\alias{weighted_prop_HImorethan5}
\title{Calculate the weighted proportion of individuals with titer more than 5}
\usage{
weighted_prop_HImorethan5(resampled_df, age_prop)
}
\arguments{
\item{resampled_df}{A dataframe containing raw titer values and age group information}

\item{age_prop}{A vector of age group proportions}
}
\value{
The weighted proportion for the population
}
\description{
This function calculates the weighted proportion of individuals with a titer value greater than or equal to 10.
}
